Split classes out into their own files.
